form update from htut

+++ src/app/features/bonusHistory/bonus-his/bonus-his.component.ts
@@ -5,6 +5,8 @@ import { MatSort } from '@angular/material/sort';
 import { BonusHistory } from 'src/app/core/models/bonus_his.model';
 import { BonusHistoryService } from 'src/app/core/services/bonus-history/bonus-history.service';
 import { ToastsService } from 'src/app/shared/toasts.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogComponent } from 'src/app/shared/dialog/dialog.component';
 @Component({
   selector: 'app-bonus-his',
   templateUrl: './bonus-his.component.html',
@@ -12,11 +14,14 @@ import { ToastsService } from 'src/app/shared/toasts.service';
 })
 export class BonusHisComponent implements OnInit {
   bonusHis: BonusHistory[] = []
-  displayedColumns: string[] = ["position","code", "bonus", "employee", "bonusDate", "remark", "amount","action"]
+  displayedColumns: string[] = ["position", "code", "bonus", "employee", "bonusDate", "remark", "amount", "action"]
   dataSource!: MatTableDataSource<BonusHistory>
   @ViewChild(MatSort, { static: true }) sort!: MatSort
 
-  constructor(private route: Router, private bonusHisService: BonusHistoryService, private toastService: ToastsService) { }
+  constructor(
+    private route: Router, private bonusHisService: BonusHistoryService,
+    private toastService: ToastsService, public dialog: MatDialog
+  ) { }
 
   ngOnInit(): void {
     this.getBonusHistory()
@@ -29,8 +34,8 @@ export class BonusHisComponent implements OnInit {
       this.dataSource.sort = this.sort
       this.dataSource.filterPredicate = (data: BonusHistory, filter: string) => {
         return data.bonusDate.toLowerCase().includes(filter) ||
-          // data.bonus.description.includes(filter) ||
-          // data.employee.description.includes(filter) ||
+          data.bonus.description.includes(filter) ||
+          data.employee.description.includes(filter) ||
           data.amount.toString().toLowerCase().includes(filter)
       }
     })
@@ -42,23 +47,28 @@ export class BonusHisComponent implements OnInit {
     this.bonusHisService._bonusHis = bonusHis;
   }
 
+  //delete bonus History
+  removeBonusHistory(bonusHis: BonusHistory) {
+    this.dialog.open(DialogComponent)
+      .afterClosed().subscribe(confirm => {
+        if (confirm) {
+          this.bonusHisService.removeBonusHistory(bonusHis.bonusHisId).subscribe(data => {
+            if (data.message == "Used") {
+              this.toastService.showWarningToast('', 'the selected ' + bonusHis.bonusHisId + ' is Used')
+              return
+            }
+            this.bonusHisService._bonus_his = this.bonusHisService._bonus_his.filter(data => data.bonusHisId != bonusHis.bonusHisId)
+            this.getBonusHistory()
+            this.toastService.showSuccessToast('', 'Success deleting ' + bonusHis.bonusHisId + '')
+          })
+        }
+      })
+  }
+
   //apply filter to table
   applyFilter(event: any) {
     let filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
-  }
-
-  //delete bonus History
-  removeBonusHistory(bonusHis:BonusHistory){
-    this.bonusHisService.removeBonusHistory(bonusHis.bonusHisId).subscribe(data=>{
-      if(data.message=="Used"){
-        this.toastService.showWarningToast('','the selected '+bonusHis.bonusHisId+' is Used')
-        return 
-      }
-      this.bonusHisService._bonus_his= this.bonusHisService._bonus_his.filter(data=>data.bonusHisId!=bonusHis.bonusHisId)
-      this.getBonusHistory()
-      this.toastService.showSuccessToast('','Success deleting '+bonusHis.bonusHisId+'')
-    })
   }
 
 }
