+++ src/app/features/departmentHistory/dept-his/dept-his.component.ts
@@ -1,10 +1,12 @@
-import { Component, OnInit,ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort } from '@angular/material/sort';
 import { Department_History } from 'src/app/core/models/dept_his.model';
 import { DepartmentHistoryService } from 'src/app/core/services/department-history/department-history.service';
 import { ToastsService } from 'src/app/shared/toasts.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogComponent } from 'src/app/shared/dialog/dialog.component';
 @Component({
   selector: 'app-dept-his',
   templateUrl: './dept-his.component.html',
@@ -12,28 +14,30 @@ import { ToastsService } from 'src/app/shared/toasts.service';
 })
 export class DeptHisComponent implements OnInit {
 
-  deptHistory:Department_History[]=[]
-  displayedColumns:string[]=["position","code","department","employee","startDate","endDate","remark","action"]
-  dataSource!:MatTableDataSource<Department_History>
-  @ViewChild(MatSort,{static:true}) sort!:MatSort
-  constructor(private deptHisService:DepartmentHistoryService,private route:Router,private toastService:ToastsService) { }
+  deptHistory: Department_History[] = []
+  displayedColumns: string[] = ["position", "code", "department", "employee", "startDate", "endDate", "remark", "action"]
+  dataSource!: MatTableDataSource<Department_History>
+  @ViewChild(MatSort, { static: true }) sort!: MatSort
+  constructor(
+    private deptHisService: DepartmentHistoryService, private route: Router,
+    private toastService: ToastsService, public dialog: MatDialog
+  ) { }
 
   ngOnInit(): void {
     this.getDepartmentHistory()
   }
 
-
   //get all department history 
-  getDepartmentHistory(){
-    this.deptHisService.getDepartHistory().subscribe(deptHistory=>{
+  getDepartmentHistory() {
+    this.deptHisService.getDepartHistory().subscribe(deptHistory => {
       console.log(deptHistory)
-      this.deptHistory=deptHistory
-      this.dataSource=new MatTableDataSource(deptHistory)
-      this.dataSource.sort=this.sort
-      this.dataSource.filterPredicate=(data:Department_History,filter:string)=>{
+      this.deptHistory = deptHistory
+      this.dataSource = new MatTableDataSource(deptHistory)
+      this.dataSource.sort = this.sort
+      this.dataSource.filterPredicate = (data: Department_History, filter: string) => {
         // return data.department.description.toLowerCase().includes(filter) ||
         //   data.employee.description.toLowerCase().includes(filter) ||
-         return  data.startDate.toLowerCase().includes(filter) ||
+        return data.startDate.toLowerCase().includes(filter) ||
           data.endDate.toLowerCase().includes(filter) ||
           data.remark.toLowerCase().includes(filter)
       }
@@ -41,29 +45,33 @@ export class DeptHisComponent implements OnInit {
   }
 
   //get department history data from row & to department history setup
-  getDeptHisDataRow(deptHis:Department_History){
+  getDeptHisDataRow(deptHis: Department_History) {
     this.route.navigate(['/main/department-assign-setup']);
-    this.deptHisService._dept_his=deptHis;
+    this.deptHisService._dept_his = deptHis;
   }
- 
+
   //apply filter to table
-  applyFilter(event:any){
-    let filterValue=(event.target as HTMLInputElement).value;
-    this.dataSource.filter=filterValue.trim().toLowerCase();
+  applyFilter(event: any) {
+    let filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
   //delete Department History
-  removeDepartmentHistory(deptHis:Department_History){
-    this.deptHisService.removeDepartmentHistory(deptHis.deptHisId).subscribe(data=>{
-      if(data.message=="Used")
-      {
-        this.toastService.showWarningToast('','the selected '+deptHis.deptHisId+' is Used')
-        return
-      }
-      this.deptHisService._dept_history= this.deptHisService._dept_history.filter(data=>data.deptHisId!=deptHis.deptHisId)
-      this.getDepartmentHistory()
-      this.toastService.showSuccessToast('','Success deleting '+deptHis.deptHisId+'')
-    })
+  removeDepartmentHistory(deptHis: Department_History) {
+    this.dialog.open(DialogComponent)
+      .afterClosed().subscribe(confirm => {
+        if (confirm) {
+          this.deptHisService.removeDepartmentHistory(deptHis.deptHisId).subscribe(data => {
+            if (data.message == "Used") {
+              this.toastService.showWarningToast('', 'the selected ' + deptHis.deptHisId + ' is Used')
+              return
+            }
+            this.deptHisService._dept_history = this.deptHisService._dept_history.filter(data => data.deptHisId != deptHis.deptHisId)
+            this.getDepartmentHistory()
+            this.toastService.showSuccessToast('', 'Success deleting ' + deptHis.deptHisId + '')
+          })
+        }
+      })
   }
 
 }
